Make installer more resilliant
This checked in file is an exact copy of:

curl -LsSfO https://astral.sh/uv/0.7.2/install.sh

Checking in the 0.7.2 version, because now a user can install with
access to github.com alone. Even if astral.sh is down for whatever
reason.

We may want to update uv installer from time to time.

Signed-off-by: Eric Curtin <ecurtin@redhat.com>